requirements.txt
Django==4.2.11
celery==5.3.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-dotenv==1.0.1
pytz
redis==5.0.3
requests==2.31.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-celery-beat
paho-mqtt